Moved find_pack_folder to tools_core
