+++ demisto_sdk/commands/common/tests/tools_core_test.py
@@ -1,6 +1,8 @@
+from pathlib import Path
+
 import pytest
 
-from demisto_sdk.commands.common.tools_core import string_to_bool
+from demisto_sdk.commands.common.tools_core import find_pack_folder, string_to_bool
 
 
 @pytest.mark.parametrize("value", ("true", "True", 1, "1", "yes", "y"))
@@ -22,3 +24,21 @@ def test_string_to_bool_false(value: str):
 def test_string_to_bool_error(value: str):
     with pytest.raises(ValueError):
         string_to_bool(value)
+
+
+@pytest.mark.parametrize(
+    "input_path,expected_output",
+    [
+        (
+            Path("root/Packs/MyPack/Integrations/MyIntegration/MyIntegration.yml"),
+            "root/Packs/MyPack",
+        ),
+        (Path("Packs/MyPack1/Scripts/MyScript/MyScript.py"), "Packs/MyPack1"),
+        (Path("Packs/MyPack2/Scripts/MyScript"), "Packs/MyPack2"),
+        (Path("Packs/MyPack3/Scripts"), "Packs/MyPack3"),
+        (Path("Packs/MyPack4"), "Packs/MyPack4"),
+    ],
+)
+def test_find_pack_folder(input_path, expected_output):
+    output = find_pack_folder(input_path)
+    assert expected_output == str(output)
