+++ demisto_sdk/commands/common/tools.py
@@ -1,4 +1,3 @@
-import contextlib
 import glob
 import io
 import os
@@ -1105,18 +1104,6 @@ def old_get_latest_release_notes_text(rn_path):
         new_rn = rn.replace(UNRELEASE_HEADER, "")  # type: ignore
 
     return new_rn if new_rn else None
-
-
-def find_pack_folder(path: Path) -> Path:
-    """
-    Finds the pack folder.
-    """
-
-    if "Packs" not in path.parts:
-        raise ValueError(f"Could not find a pack for {str(path)}")
-    if path.parent.name == "Packs":
-        return path
-    return path.parents[len(path.parts) - (path.parts.index("Packs")) - 3]
 
 
 def get_release_notes_file_path(file_path):
