+++ demisto_sdk/commands/init/initiator.py
@@ -58,13 +58,13 @@ from demisto_sdk.commands.common.handlers import DEFAULT_JSON_HANDLER as json
 from demisto_sdk.commands.common.handlers import DEFAULT_YAML_HANDLER as yaml
 from demisto_sdk.commands.common.logger import logger
 from demisto_sdk.commands.common.tools import (
-    find_pack_folder,
     get_common_server_path,
     get_file,
     get_pack_name,
     get_yaml,
 )
 from demisto_sdk.commands.common.tools_core import (
+    find_pack_folder,
     is_external_repository,
     string_to_bool,
 )
